Made it compile and run on windows again.. It compiles with 1146 Wearnings and the program segfaults afeter few seconds, but it works.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@499 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Utility/GameState.cpp
@@ -585,7 +585,7 @@ void GameState::initAsteroids() {
    double asteroidSize;
    // Spawn the initial asteroids for the game.
    for (int i = 0; i < numAsteroidsToSpawn; ++i) {
-      asteroidSize = (minAsteroidSize + asteroidSizeVariation * randdouble()) / sqrt(numAsteroidsToSpawn);
+      asteroidSize = (minAsteroidSize + asteroidSizeVariation * randdouble()) / sqrt((double)numAsteroidsToSpawn);
       tempAsteroid = new Asteroid3D(minAsteroidSize + asteroidSizeVariation * randdouble(), worldSize);
       // Add each asteroid to the custodian so we know of its existence.
       custodian.add(tempAsteroid);
